requirements.txt
feedparser
aiosqlite
beautifulsoup4
python-dotenv
aiogram